chore(e2e-test): split end to end test
